+++ src/internals/tact/index.ts
@@ -3,3 +3,4 @@ export * from "./util";
 export * from "./constEval";
 export * from "./stdlib";
 export * from "./config";
+export * from "./parser";